LayoutTests/platform/qt/svg/dom/font-face-elements-expected.txt

This tests to make sure we have all the font-face related dom bindings.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS font-face in http://www.w3.org/2000/svg is JS object SVGFontFaceElement
PASS font-face-name in http://www.w3.org/2000/svg is JS object SVGFontFaceNameElement
PASS font-face-format in http://www.w3.org/2000/svg is JS object SVGFontFaceFormatElement
PASS font-face-src in http://www.w3.org/2000/svg is JS object SVGFontFaceSrcElement
PASS font-face-uri in http://www.w3.org/2000/svg is JS object SVGFontFaceUriElement
PASS successfullyParsed is true

TEST COMPLETE
